mcp_servers/memory_storage_mcp/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.7.0
python-multipart==0.0.6
pyyaml>=6.0.1
httpx>=0.27.0
aiofiles==23.1.0
python-dotenv>=1.0.1
tenacity==8.2.2
markdown==3.4.3
schedule>=1.2.1
pytest>=8.0.0
pytest-cov>=6.0.0
mypy>=1.5.0
pytest-xdist>=3.5.0
pytest-timeout>=2.2.0
pytest-mock>=3.10.0 